inline find input component
