+++ src/vs/workbench/contrib/positronNotebook/browser/contrib/find/PositronFindWidget.tsx
@@ -7,14 +7,15 @@
 import './FindWidget.css';
 
 // React.
-import React from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 
 // Other dependencies.
 import { ActionButton } from '../../utilityComponents/ActionButton.js';
-import { PositronFindInput } from './PositronFindInput.js';
+import { FindInput, IFindInputOptions } from '../../../../../../base/browser/ui/findinput/findInput.js';
 import { Toggle } from '../../../../../../base/browser/ui/toggle/toggle.js';
 import { ISettableObservable } from '../../../../../../base/common/observable.js';
 import { useObservedValue } from '../../useObservedValue.js';
+import { defaultInputBoxStyles, defaultToggleStyles } from '../../../../../../platform/theme/browser/defaultStyles.js';
 
 export interface PositronFindWidgetProps {
 	readonly findText: ISettableObservable<string>;
@@ -25,6 +26,7 @@ export interface PositronFindWidgetProps {
 	readonly matchIndex?: number;
 	readonly matchCount?: number;
 	readonly additionalToggles?: Toggle[];
+	readonly findInputOptions?: IFindInputOptions;
 	readonly onPreviousMatch: () => void;
 	readonly onNextMatch: () => void;
 	readonly onClose: () => void;
@@ -39,6 +41,7 @@ export const PositronFindWidget = ({
 	matchIndex,
 	matchCount,
 	additionalToggles,
+	findInputOptions,
 	onPreviousMatch,
 	onNextMatch,
 	onClose,
@@ -48,23 +51,119 @@ export const PositronFindWidget = ({
 	const matchWholeWord = useObservedValue(matchWholeWordObs);
 	const useRegex = useObservedValue(useRegexObs);
 
+	const findInputContainerRef = useRef<HTMLDivElement>(null);
+	const [findInput, setFindInput] = useState<FindInput | null>(null);
+
+	// Capture initial options to avoid recreating FindInput on prop changes
+	const initialOptionsRef = useRef({
+		...findInputOptions,
+		additionalToggles,
+	});
+
+	// Initialize FindInput widget once on mount
+	useEffect(() => {
+		if (!findInputContainerRef.current) {
+			return;
+		}
+
+		const options = initialOptionsRef.current;
+
+		// Create the FindInput widget with merged options
+		const input = new FindInput(
+			findInputContainerRef.current,
+			undefined,  // context view provider
+			{
+				label: 'Find',
+				placeholder: 'Find',
+				showCommonFindToggles: true,
+				...options,
+				inputBoxStyles: defaultInputBoxStyles,
+				toggleStyles: defaultToggleStyles,
+			});
+
+		setFindInput(input);
+
+		return () => {
+			input.dispose();
+			setFindInput(null);
+		};
+	}, []);
+
+	// Set up onInput listener
+	useEffect(() => {
+		if (!findInput) {
+			return;
+		}
+
+		const disposable = findInput.onInput(() => {
+			findTextObs.set(findInput.getValue(), undefined);
+		});
+
+		return () => disposable.dispose();
+	}, [findInput, findTextObs]);
+
+	// Set up onDidOptionChange listener
+	useEffect(() => {
+		if (!findInput) {
+			return;
+		}
+
+		const disposable = findInput.onDidOptionChange(() => {
+			matchCaseObs.set(findInput.getCaseSensitive(), undefined);
+			matchWholeWordObs.set(findInput.getWholeWords(), undefined);
+			useRegexObs.set(findInput.getRegex(), undefined);
+		});
+
+		return () => disposable.dispose();
+	}, [findInput, matchCaseObs, matchWholeWordObs, useRegexObs]);
+
+	// Update additionalToggles dynamically
+	useEffect(() => {
+		if (findInput && additionalToggles) {
+			findInput.setAdditionalToggles(additionalToggles);
+		}
+	}, [findInput, additionalToggles]);
+
+	// Update value
+	useEffect(() => {
+		if (findInput && findInput.getValue() !== findText) {
+			findInput.setValue(findText || '');
+		}
+	}, [findInput, findText]);
+
+	// Update toggle states
+	useEffect(() => {
+		if (findInput) {
+			findInput.setCaseSensitive(matchCase);
+		}
+	}, [findInput, matchCase]);
+
+	useEffect(() => {
+		if (findInput) {
+			findInput.setWholeWords(matchWholeWord);
+		}
+	}, [findInput, matchWholeWord]);
+
+	useEffect(() => {
+		if (findInput) {
+			findInput.setRegex(useRegex);
+		}
+	}, [findInput, useRegex]);
+
+	// Focus input when requested
+	useEffect(() => {
+		if (focusInput && findInput) {
+			findInput.focus();
+			findInput.select();
+		}
+	}, [findInput, focusInput]);
+
 	return (
 		<div className='positron-find-widget-positioned'>
 			<div className='positron-find-widget'>
 				<div className='find-widget-row'>
 					<div className='find-input-container'>
-						<PositronFindInput
-							additionalToggles={additionalToggles}
-							focusInput={focusInput}
-							matchCase={matchCase}
-							matchWholeWord={matchWholeWord}
-							useRegex={useRegex}
-							value={findText}
-							onMatchCaseChange={(value) => matchCaseObs.set(value, undefined)}
-							onMatchWholeWordChange={(value) => matchWholeWordObs.set(value, undefined)}
-							onUseRegexChange={(value) => useRegexObs.set(value, undefined)}
-							onValueChange={(value) => findTextObs.set(value, undefined)}
-						/>
+						<div ref={findInputContainerRef} />
 					</div>
 					<div className={`find-results ${findText && matchCount === 0 ? 'no-results' : ''}`}>
 						{findText && matchCount !== undefined ? (
